[MWPW-136595] Move critical FEDS block checks to Milo (#180)
* [MWPW-136595] Removed whitespaces (eslint warnings)

* [MWPW-136595] Moved FEDS critical checks to Milo folder

* [MWPW-136595] Moved FEDS critical checks to Milo folder (II)

* [MWPW-136595] Force GDPR consent on US servers

* [MWPW-136595] Stabilize FEDS checks on Milo (I)

* [MWPW-136595] Moved Jarvis checks back to FEDS project

* [MWPW-136595] Cleanup PR comments / Re-trigger RunNala

* [MWPW-136595] Commenting Consent test step to avoid flakiness

* [MWPW-136595] Removed Consent test step to avoid flakiness (II)

---------
