+++ features/milo/footer.spec.js
@@ -7,9 +7,9 @@ module.exports = {
         '/libs/feds/drafts/qa/footer/feds-default-footer',
       ],
       wcagTags: ['wcag2a', 'wcag2aa', 'wwcag21a', 'wcag21aa'],
-      browserParams: '?georouting=off',
+      browserParams: '?customPrivacyLocation=de&georouting=off',
       envs: '@milo_live',
-      tags: '@feds @footer @feds-footer',
+      tags: '@milo @feds @footer @smoke @regression',
     },
     {
       name: '@FEDS-Skinny-Footer',
@@ -17,9 +17,9 @@ module.exports = {
         '/libs/feds/drafts/qa/footer/feds-skinny-footer',
       ],
       wcagTags: ['wcag2a', 'wcag2aa', 'wwcag21a', 'wcag21aa'],
-      browserParams: '?georouting=off',
+      browserParams: '?customPrivacyLocation=de&georouting=off',
       envs: '@milo_live',
-      tags: '@feds @footer @feds-footer',
+      tags: '@milo @feds @footer @smoke @regression',
     },
     {
       name: '@FEDS-Privacy-Footer',
@@ -27,19 +27,9 @@ module.exports = {
         '/libs/feds/drafts/qa/footer/feds-privacy-footer',
       ],
       wcagTags: ['wcag2a', 'wcag2aa', 'wwcag21a', 'wcag21aa'],
-      browserParams: '?georouting=off',
+      browserParams: '?customPrivacyLocation=de&georouting=off',
       envs: '@milo_live',
-      tags: '@feds @footer @feds-footer',
-    },
-    {
-      name: '@DC-Footer-Checks',
-      path: [
-        '/acrobat/online/sign-pdf.html',
-      ],
-      wcagTags: ['wcag2a', 'wcag2aa', 'wwcag21a', 'wcag21aa'],
-      browserParams: '?georouting=off',
-      envs: '@adobe_prod',
-      tags: '@dc @footer @dc-footer',
+      tags: '@milo @feds @footer @smoke @regression',
     },
   ],
 };
