+++ selectors/blog/404.page.js
@@ -1,9 +1,9 @@
 export default class PageNotFound {
   constructor(page) {
     this.page = page;
-       
+
     // 404 fragment
     this.fragment = this.page.locator('.fragment').nth(0);
-    this.header1Text = this.fragment.locator('#these-are-uncharted-waters');   
+    this.header1Text = this.fragment.locator('#these-are-uncharted-waters');
   }
 }
