+++ tests/milo/breadcrumbs.test.js
@@ -2,7 +2,7 @@
 /* eslint-disable no-await-in-loop */
 /* eslint-disable import/extensions */
 import { expect, test } from '@playwright/test';
-import { features } from '../../features/feds/breadcrumbs.spec.js';
+import { features } from '../../features/milo/breadcrumbs.spec.js';
 import FedsHeader from '../../selectors/feds/feds.header.page.js';
 
 test.describe('Breadcrumbs Component test suite', () => {
